+++ app/src/components/market/market_wizard_creator.tsx
@@ -10,14 +10,7 @@ import { Arbitrator, MarketData, Question, Token } from '../../util/types'
 import { BigNumberInputReturn } from '../common/big_number_input'
 
 import { Outcome } from './outcomes'
-import {
-  AskQuestionStep,
-  CreateMarketStep,
-  FundingAndFeeStep,
-  MenuStep,
-  OutcomesStep,
-  SummaryMarketStep,
-} from './steps'
+import { AskQuestionStep, CreateMarketStep, FundingAndFeeStep, MenuStep, SummaryMarketStep } from './steps'
 
 interface Props {
   callback: (param: MarketData) => void
@@ -240,15 +233,17 @@ export const MarketWizardCreator = (props: Props) => {
             handleClearQuestion={handleClearQuestion}
             handleDateChange={handleDateChange}
             handleQuestionChange={handleQuestionChange}
+            handleOutcomesChange={handleOutcomesChange}
             next={() => next()}
             values={{
               question,
+              outcomes,
+              loadedQuestionId,
               category,
               categoriesCustom,
               resolution,
               arbitrator,
               arbitratorsCustom,
-              loadedQuestionId,
             }}
           />
         )
@@ -260,19 +255,6 @@ export const MarketWizardCreator = (props: Props) => {
             handleCollateralChange={handleCollateralChange}
             next={() => next()}
             values={{ collateral, spread, funding }}
-          />
-        )
-      case 3:
-        return (
-          <OutcomesStep
-            back={() => back()}
-            handleOutcomesChange={handleOutcomesChange}
-            next={() => next()}
-            values={{
-              outcomes,
-              question,
-              loadedQuestionId,
-            }}
           />
         )
       case 4:
@@ -294,15 +276,17 @@ export const MarketWizardCreator = (props: Props) => {
             handleClearQuestion={handleClearQuestion}
             handleDateChange={handleDateChange}
             handleQuestionChange={handleQuestionChange}
+            handleOutcomesChange={handleOutcomesChange}
             next={() => next()}
             values={{
               question,
+              outcomes,
+              loadedQuestionId,
               category,
               categoriesCustom,
               resolution,
               arbitrator,
               arbitratorsCustom,
-              loadedQuestionId,
             }}
           />
         )
