+++ app/src/components/market/steps/items/ask_question_step.tsx
@@ -7,6 +7,7 @@ import {
   Arbitrators,
   Button,
   ButtonContainer,
+  ButtonLink,
   Categories,
   CreateCard,
   DateField,
@@ -15,6 +16,9 @@ import {
   Well,
 } from '../../../common'
 import { QuestionInput } from '../../../common/question_input'
+
+import { MAX_OUTCOME_ALLOWED } from '../../../../common/constants'
+import { Outcome, Outcomes } from '../../outcomes'
 
 interface Props {
   next: () => void
@@ -26,11 +30,13 @@ interface Props {
     arbitrator: Arbitrator
     arbitratorsCustom: Arbitrator[]
     loadedQuestionId: Maybe<string>
+    outcomes: Outcome[]
   }
   handleChange: (event: ChangeEvent<HTMLInputElement> | ChangeEvent<HTMLSelectElement>) => any
   handleDateChange: (date: Date | null) => any
   handleQuestionChange: (question: Question, arbitrator: Arbitrator) => any
   handleArbitratorChange: (arbitrator: Arbitrator) => any
+  handleOutcomesChange: (newOutcomes: Outcome[]) => any
   handleClearQuestion: () => any
   addArbitratorCustom: (arbitrator: Arbitrator) => void
   addCategoryCustom: (category: string) => void
@@ -39,6 +45,17 @@ interface Props {
 const OracleInfo = styled(Well)`
   font-style: italic;
   text-align: center;
+`
+const ButtonLinkStyled = styled(ButtonLink)`
+  margin-right: auto;
+`
+
+const OutcomeInfo = styled(Well)`
+  margin-bottom: 30px;
+`
+
+const ButtonContainerStyled = styled(ButtonContainer)`
+  padding-top: 10px;
 `
 
 const AskQuestionStep = (props: Props) => {
@@ -52,12 +69,21 @@ const AskQuestionStep = (props: Props) => {
     handleClearQuestion,
     handleDateChange,
     handleQuestionChange,
+    handleOutcomesChange,
     next,
     values,
   } = props
-  const { arbitrator, arbitratorsCustom, categoriesCustom, category, loadedQuestionId, question, resolution } = values
 
-  const error = !question || !category || !resolution
+  const {
+    arbitrator,
+    arbitratorsCustom,
+    categoriesCustom,
+    category,
+    loadedQuestionId,
+    outcomes,
+    question,
+    resolution,
+  } = values
 
   const validate = (e: any) => {
     e.preventDefault()
@@ -66,6 +92,37 @@ const AskQuestionStep = (props: Props) => {
       next()
     }
   }
+  const errorMessages = []
+
+  const someEmptyName = outcomes.some(outcome => !outcome.name)
+  if (someEmptyName) {
+    errorMessages.push('The names of the outcomes should not be empty.')
+  }
+
+  const someEmptyProbability = outcomes.some(outcome => outcome.probability === 0)
+  if (someEmptyProbability) {
+    errorMessages.push('The probabilities of the outcomes should not be zero.')
+  }
+
+  const totalProbabilities = outcomes.reduce((total, cur) => total + cur.probability, 0)
+  if (totalProbabilities !== 100) {
+    errorMessages.push('The sum of all probabilities must be equal to 100%.')
+  }
+
+  if (outcomes.length < 2) {
+    errorMessages.push('Please enter at least 2 outcomes')
+  }
+
+  const error =
+    totalProbabilities !== 100 ||
+    someEmptyName ||
+    someEmptyProbability ||
+    outcomes.length < 2 ||
+    !question ||
+    !category ||
+    !resolution
+
+  const canAddOutcome = outcomes.length < MAX_OUTCOME_ALLOWED && !loadedQuestionId
 
   return (
     <CreateCard>
@@ -90,6 +147,19 @@ const AskQuestionStep = (props: Props) => {
           description: `Ensure that your market is phrased in a way that it can be unambiguous, resolvable, contains all the necessary parameters and is clearly understandable. As a guide you could use the <a target="_blank" href="https://augur.guide/2-market-creators/checksheet.html">check sheet</a> prepared by Augur for creating markets.`,
         }}
       />
+
+      <OutcomeInfo>
+        Please add all the possible outcomes for the <strong>&quot;{question}&quot;</strong> question.
+      </OutcomeInfo>
+      <Outcomes
+        canAddOutcome={canAddOutcome}
+        disabled={!!loadedQuestionId}
+        errorMessages={errorMessages}
+        onChange={handleOutcomesChange}
+        outcomes={outcomes}
+        totalProbabilities={totalProbabilities}
+      />
+
       <FormRow
         formField={
           <Categories
@@ -143,11 +213,12 @@ const AskQuestionStep = (props: Props) => {
       <OracleInfo>
         <DisplayArbitrator arbitrator={arbitrator} />
       </OracleInfo>
-      <ButtonContainer>
-        <Button disabled={error} onClick={validate}>
+      <ButtonContainerStyled>
+        <ButtonLinkStyled onClick={props.next}>‹ Cancel</ButtonLinkStyled>
+        <Button disabled={error} onClick={props.next}>
           Next
         </Button>
-      </ButtonContainer>
+      </ButtonContainerStyled>
     </CreateCard>
   )
 }
